+++ projects/shared/src/lib/layout/layout.css
@@ -24,3 +24,17 @@
   padding: 16px;
 }
 
+.app-footer {
+  margin-top: auto;
+  padding: 8px 16px;
+  background: var(--md-sys-color-surface);
+  color: var(--md-sys-color-on-surface);
+  border-top: 1px solid var(--md-sys-color-outline-variant);
+}
+
+.layout-header-slot,
+.layout-footer-slot {
+  max-width: 1200px;
+  margin: 0 auto;
+}
+
