+++ projects/shared/src/lib/layout/layout.html
@@ -1,6 +1,11 @@
 <div class="app-layout">
-  <header class="app-header"><shared-header></shared-header></header>
-  
+  <header class="app-header">
+    <shared-header></shared-header>
+    <div class="layout-header-slot">
+      <ng-content select="[layout-header]"></ng-content>
+    </div>
+  </header>
+
   <mat-sidenav-container class="sidenav-container">
     <mat-sidenav-content class="main-content">
       <main class="app-main">
@@ -8,4 +13,10 @@
       </main>
     </mat-sidenav-content>
   </mat-sidenav-container>
+
+  <footer class="app-footer">
+    <div class="layout-footer-slot">
+      <ng-content select="[layout-footer]"></ng-content>
+    </div>
+  </footer>
 </div>
