+++ projects/shared/src/lib/step-controls/step-controls.css
@@ -1,0 +1,6 @@
+.step-controls {
+  display: flex;
+  justify-content: space-between;
+  gap: 0.75rem;
+}
+
