+++ projects/shared/src/lib/step-controls/step-controls.html
@@ -1,0 +1,11 @@
+<div class="step-controls">
+  <button mat-stroked-button color="primary" (click)="navigatePrev()" [disabled]="!prevUrl()">
+    <mat-icon>arrow_back</mat-icon>
+    Previous
+  </button>
+  <button mat-raised-button color="primary" (click)="navigateNext()" [disabled]="!nextUrl()">
+    Next
+    <mat-icon>arrow_forward</mat-icon>
+  </button>
+</div>
+
