+++ projects/shared/src/lib/step-layout/step-layout.html
@@ -1,18 +1,12 @@
 <section class="step-layout">
-  <shared-step-navigation></shared-step-navigation>
+  <div layout-header>
+    <shared-step-navigation></shared-step-navigation>
+  </div>
   <div class="step-content">
     <router-outlet></router-outlet>
   </div>
-  <div class="step-nav-actions">
-    <button mat-stroked-button color="primary" (click)="navigatePrev()" [disabled]="!prevUrl()">
-      <mat-icon>arrow_back</mat-icon>
-      Previous
-    </button>
-    <button mat-raised-button color="primary" (click)="navigateNext()" [disabled]="!nextUrl()">
-      Next
-      <mat-icon>arrow_forward</mat-icon>
-    </button>
+  <div layout-footer>
+    <shared-step-controls></shared-step-controls>
   </div>
-  
 </section>
